// src/api/generated.ts
export type GetApiServerinfo200Memory = { [key: string]: unknown };

export type GetApiServerinfo200Cpu = { [key: string]: unknown };

export type GetApiServerinfo200 = {
  platform?: string;
  uptime?: string;
  memory?: GetApiServerinfo200Memory;
  cpu?: GetApiServerinfo200Cpu;
};